requirements.txt
llama-index==0.14.21
llama-index-llms-google-genai==0.9.1
llama-index-embeddings-google-genai==0.5.0
python-dotenv==1.2.2
pandas==3.0.2
